refactor: simplify per /simplify pass
- Extract the duplicated relationId helper (verify.ts + recipients.ts) to
  a shared src/lib/utilities/relationId.ts.
- notifications/types.ts: top-level ReminderLevel import instead of an
  inline import() type.
- recipients.ts: reuse the already-populated event.region.breadcrumbs (the
  job loads events at depth 1) instead of an extra findByID per dispatch.

Refs #484
